fix: load tsconfig.json in official TypeScript way
